refactored profile page
